% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idoveData.R
\name{idoveData}
\alias{idoveData}
\title{Toy Dataset For Illustration}
\format{
idoveData is a data.frame containing 40,000 participants. 
  The data.frame contains 6 columns, 
  \describe{
  \item{entry.time}{The entry time in days.}
  \item{left.time}{The left end of the time interval in days.}
  \item{right.time}{The right end of the time interval in days.}
  \item{vaccine.time}{The time of vaccination in days.}
  \item{priority}{A composite baseline risk score taking values 1-5.}
  \item{sex}{A binary indicator of sex (male/female).}
  }
}
\usage{
data(idoveData)
}
\description{
This data set is provided for the purposes of illustrating the use of
the software. It was simulated under a blinded, priority-tier dependent 
crossover design.
}
\keyword{datasets}
