% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{intCens}
\alias{intCens}
\title{Specify Time Variables}
\usage{
intCens(entry_time, left_time, right_time, vaccination_time)
}
\arguments{
\item{entry_time}{The variable for the time when 
the participant enters the trial.}

\item{left_time}{The variable for the last examination time when
the test is negative}

\item{right_time}{The variable for the first examination time when
the test is positive; Inf or NA if the participant is never tested
positive during the follow-up}

\item{vaccination_time}{The variable for the time when 
vaccination takes place.}
}
\value{
This function is intended to be used only in the model statement 
 of idove(). The result, a matrix, is used internally.
}
\description{
This function is used in the model statement of idove() to specify
 the entry time, left interval time, right interval time, and vaccination time.
}
