% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.iDOVE.R
\name{print}
\alias{print}
\alias{print.iDOVE}
\title{Print the Primary Results of an idove() Analysis}
\usage{
\method{print}{iDOVE}(x, ...)
}
\arguments{
\item{x}{An iDOVE object. The value object returned by a call to idove()}

\item{...}{ignored}
}
\value{
No return value, called to display key results.
}
\description{
Print the primary results of an idove() analysis.
}
\examples{

data(idoveData)

set.seed(1234)
smp <- sample(1L:nrow(x = idoveData), size = 250L)

# NOTE: This sample size is chosen for example only -- larger data sets
# should be used.
# See the vignette for a full analysis of the idoveData dataset

# Fit the model with default settings
result <- idove(formula = intCens(entry.time, left.time, right.time, vaccine.time) ~ 1, 
                data = idoveData[smp,])

print(x = result)

}
